+++ ci-runners-ts/index.ts
@@ -1,6 +1,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as aws from "@pulumi/aws";
 import { CodebuildProjectArgs, createCodebuildProject } from "./codebuild-project";
+import { createCodebuildRole } from "./iam";
 
 const config = new pulumi.Config();
 // The GitHub repository where the codebuild project will be used.
@@ -11,37 +12,9 @@ const githubConnection = new aws.codeconnections.Connection("github_connection",
     name: codeConnectionName,
     providerType: "GitHub",
 });
-// Grant CodeBuild project IAM role access to use the connection, as documented in
-// https://docs.aws.amazon.com/codebuild/latest/userguide/connections-github-app.html#connections-github-role-access
-const codebuildPolicyDoc = aws.iam.getPolicyDocumentOutput({
-    statements: [{
-        effect: "Allow",
-        principals: [{
-            type: "Service",
-            identifiers: ["codebuild.amazonaws.com"],
-        }],
-        actions: ["sts:AssumeRole"],
-    }],
-});
-const codebuildRole = new aws.iam.Role("codebuild_role", {
-    name: "codebuild-github-runner-role",
-    assumeRolePolicy: codebuildPolicyDoc.apply(codebuildPolicyDoc => codebuildPolicyDoc.json),
-});
-const codebuildPolicy = new aws.iam.RolePolicy("codebuild_policy", {
-    name: "codebuild-github-runner-policy",
-    role: codebuildRole.id,
-    policy: pulumi.jsonStringify({
-        Version: "2012-10-17",
-        Statement: [{
-            Effect: "Allow",
-            Action: [
-                "codeconnections:GetConnectionToken",
-                "codeconnections:GetConnection",
-            ],
-            Resource: [githubConnection.arn],
-        }],
-    }),
-});
+
+// Create the CodeBuild IAM role with the GitHub connection
+const codebuildRole = createCodebuildRole(githubConnection.arn);
 
 createCodebuildProject("ubuntu222c", {
     name: "ubuntu-22-2c",
